// SDKs/googleapis.com/displayvideo/v1/typescript/src/sdk/models/shared/campaigngoal.ts
import { SpeakeasyMetadata, SpeakeasyBase } from "../../../internal/utils";
import { PerformanceGoal } from "./performancegoal";


export enum CampaignGoalCampaignGoalTypeEnum {
    CampaignGoalTypeUnspecified = "CAMPAIGN_GOAL_TYPE_UNSPECIFIED",
    CampaignGoalTypeAppInstall = "CAMPAIGN_GOAL_TYPE_APP_INSTALL",
    CampaignGoalTypeBrandAwareness = "CAMPAIGN_GOAL_TYPE_BRAND_AWARENESS",
    CampaignGoalTypeOfflineAction = "CAMPAIGN_GOAL_TYPE_OFFLINE_ACTION",
    CampaignGoalTypeOnlineAction = "CAMPAIGN_GOAL_TYPE_ONLINE_ACTION"
}


// CampaignGoal
/** 
 * Settings that control the goal of a campaign.
**/
export class CampaignGoal extends SpeakeasyBase {
  @SpeakeasyMetadata({ data: "json, name=campaignGoalType" })
  campaignGoalType?: CampaignGoalCampaignGoalTypeEnum;

  @SpeakeasyMetadata({ data: "json, name=performanceGoal" })
  performanceGoal?: PerformanceGoal;
}